move integration test utilities to Util module.
